feat: Fix SBY pathing bug (Flaw 5) and logic optimization gap (Flaw 2)
- Auto-link .sv and .v files during compilation via Workspace loaded files
- Add abc logic optimization pass and --optimize CLI flag
- Fix minor clippy warnings in pipeline

+++ src/emit/rspu_encoding/decode.rs
@@ -49,7 +49,7 @@ pub fn decode(word: u64, target: &TargetSpec) -> Result<RspuInstruction, MirrErr
         }
         OP_SR_INIT => {
             let (guard, cond, imm) = extract_i_fields(word, target);
-            Ok(RspuInstruction::SrInit { guard: guard as u8, length: imm as u32, cond })
+            Ok(RspuInstruction::SrInit { guard: guard as u8, length: imm, cond })
         }
         OP_SR_TICK => {
             let (guard, _, _, _) = extract_g_fields(word, target);
